end db connections
